backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
asyncpg==0.29.0

# ML/AI
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.26.3
scikit-learn==1.4.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
